Replace docker-incompatible syscall
getlogin_r (equivalent of who command) does not work well with container images because there is no logged in user. The OS will still set a real and effective uid so we can still work out the user (equivalent of whoami). These two implementations are not quite functionally equivalent but it is very close.